// libs/shared/src/models/api-models.ts
export interface LoginViewModel {
  userName: string;
  password: string;
  deviceCode?: string | null;
}